refactor: dedup splitLine and refresh stale docs (B7, B8)
- Extract the canonical line tokenizer to src/sysex-split.js (zero-dep leaf)
  and import it from parser.js and tests/helpers/sysex-fixture.js, so the
  parser and its fixture decoder can no longer diverge. The CommonJS
  capture-fixtures.cjs build tool keeps a hand-synced copy (ESM/CJS boundary;
  not in the test/runtime path) with a pointer comment.
- parser.js: correct the stale comment (isLoadingPreset clear lives in
  event-bridge.js, not main.js).
- CLAUDE.md: refresh the architecture diagram and data flow for the
  post-refactor module set; replace the "refactor in flight" cycle list with
  the completed-refactor status; toggleDspKey now lives only in navigation.js.

No behavior change; 66 tests pass.
